+ handling MSVC warnings

+++ common/source/counter.cpp
@@ -11,14 +11,14 @@
         {
             LARGE_INTEGER counter;
             QueryPerformanceCounter(&counter);
-            return counter.QuadPart;
+            return static_cast< uint64_t >( counter.QuadPart );
         }
 
         uint64_t GetCounterFrequency()
         {
             LARGE_INTEGER freq;
             QueryPerformanceFrequency(&freq);
-            return freq.QuadPart;
+            return static_cast< uint64_t >( freq.QuadPart );
         }
 
     } // namespace Common
